skills/claytantor/pylinter-assist/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pylinter-assist"
version = "0.2.0+0"
description = "Context-aware Python linting with smart pattern heuristics for PR review"
requires-python = ">=3.10"
dependencies = [
    "pylint>=3.0",
    "pyyaml>=6.0",
    "pygithub>=2.1",
    "requests>=2.31",
    "click>=8.1",
    "rich>=13.0",
]

[project.scripts]
lint-pr = "pylinter_assist.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pylinter_assist*"]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"
